+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 1.2.1:
   - make advanced scheduler the default
+  - do not process empty sync committee messages
 
 1.2.0:
   - fetch attestation duties for next epoch approximately half way through current epoch